requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
pydantic-ai-slim[openai,mcp]
mcp
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn